ki-kumpel/requirements.txt
pystray
pillow
mss
openai
tk
pyinstaller
pywin32
pyttsx3
distro
httpx
